fix(bot): Improved word selection for knowledge testing v1.0.2
The handle_quiz function in the algorithm has been modified to obtain words for testing knowledge. Instead of using all the words in the Word model, only common words and those added by the current user will be written to the words variable. This change allows for more accurate word selection and improves the logic of the bot. Additionally, this modification excludes words passed from one user to another from being included in the selection process.