refactor(packages): Reorganized package structure
The packages "fs_tools", "config", and "read_file" have been merged into one package in order to simplify the project's structure and improve the organization of the code. All functions related to working with files and configuration data are now in one location. Additionally, the "json2db" module has been relocated to the "db" package.